requirements.txt
Django==5.2.3
django-guardian==3.0.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0

